Add dev tools, separate routes
