+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "build": "tsc",
     "start": "node ./src/server.js",
-    "dev": "npm run build && npm run start"
+    "predev": "tsc",
+    "dev": "concurrently \"tsc -w\" \"nodemon ./src/server.js\""
   },
   "repository": {
     "type": "git",
@@ -34,6 +35,8 @@
     "@types/dotenv": "^8.2.0",
     "@types/express": "^4.17.7",
     "@types/helmet": "0.0.48",
-    "@types/mongoose": "^5.7.36"
+    "@types/mongoose": "^5.7.36",
+    "concurrently": "^5.3.0",
+    "nodemon": "^2.0.4"
   }
 }
